requirements.txt
streamlit>=1.33
transformers>=4.41
torch>=2.1
sentencepiece>=0.2.0
yake>=0.4.8
pypdf>=4.2
python-docx>=1.1
